About time I updated my repo

+++ LibC/src/lemon/filesystem.c
@@ -1,10 +1,10 @@
-#include <core/syscall.h>
+#include <lemon/syscall.h>
 #include <stddef.h>
-#include <core/filesystem.h>
+#include <lemon/filesystem.h>
 
 int lemon_open(char* filename, int flags){
     int fd;
-    syscall(SYS_OPEN, (uint32_t)filename, (uint32_t)&fd, 0, 0, 0);
+    syscall(SYS_OPEN, filename, &fd, 0, 0, 0);
     return fd;
 }
 
@@ -14,25 +14,25 @@ void lemon_close(int fd){
 
 int lemon_read(int fd, void* buffer, size_t count){
     int ret;
-    syscall(SYS_READ, fd, buffer, count, (uint32_t)&ret, 0);
+    syscall(SYS_READ, fd, buffer, count, &ret, 0);
     return ret;
 }
 
 int lemon_write(int fd, void* buffer, size_t count){
     int ret;
-    syscall(SYS_WRITE, fd, buffer, count, (uint32_t)&ret, 0);
+    syscall(SYS_WRITE, fd, buffer, count, &ret, 0);
     return ret;
 }
 
 int lemon_seek(int fd, uint32_t offset, int whence){
     uint32_t ret;
-    syscall(SYS_LSEEK, fd, offset, whence, (uint32_t)&ret, 0);
+    syscall(SYS_LSEEK, fd, offset, whence, &ret, 0);
     return ret;
 }
 
 int lemon_readdir(int fd, uint32_t count, lemon_dirent_t* dirent){
     int ret;
-    syscall(SYS_READDIR, fd, dirent, count, (uint32_t)&ret, 0);
+    syscall(SYS_READDIR, fd, dirent, count, &ret, 0);
     return ret;
 }
